+++ enabler/src/components/payment-methods/card/index.ts
@@ -1,4 +1,4 @@
-import { Card } from "./CardREMOVE";
+import { Card } from "./Card";
 import { CardBuilder } from "./CardBuilder";
 
 export { Card, CardBuilder };
