+++ enabler/src/components/payment-methods/invoice/index.ts
@@ -1,4 +1,4 @@
-import { Invoice } from "./InvoiceREMOVE";
+import { Invoice } from "./Invoice";
 import { InvoiceBuilder } from "./InvoiceBuilder";
 
 export { Invoice, InvoiceBuilder };
